+++ comics/comics/boxerhockey.py
@@ -9,15 +9,14 @@ class ComicData(ComicDataBase):
     rights = 'Tyson "Rittz" Hesse'
 
 class Crawler(CrawlerBase):
-    history_capable_days = 30
+    history_capable_days = 120
     time_zone = -7
-    history_capable_days = 30
 
     def crawl(self, pub_date):
         feed = self.parse_feed(
             'http://boxerhockey.fireball20xl.com/inc/feed.php')
         for entry in feed.for_date(pub_date):
             title = entry.title
-            page = self.parse_page(entry.link)
+            page = self.parse_page(entry.id)
             url = page.src('img#comicimg')
             return CrawlerImage(url, title)