+++ comics/comics/havet.py
@@ -16,6 +16,6 @@ class Crawler(CrawlerBase):
     time_zone = 1
 
     def crawl(self, pub_date):
-        url = 'http://havet.nettserier.no/_striper/havet-%s.jpg' % (
-            self.date_to_epoch(pub_date),)
+        epoch = self.date_to_epoch(pub_date, 'Europe/Oslo')
+        url = 'http://havet.nettserier.no/_striper/havet-%s.jpg' % epoch
         return CrawlerImage(url)