+++ comics/comics/rocky.py
@@ -14,6 +14,7 @@ class Crawler(CrawlerBase):
     time_zone = 1
 
     def crawl(self, pub_date):
+        epoch = self.date_to_epoch(pub_date, 'Europe/Oslo')
         url = 'http://www.dagbladet.no/tegneserie/rockyarkiv/serve.php?%s' % (
-            self.date_to_epoch(pub_date),)
+            epoch,)
         return CrawlerImage(url)