+++ comics/comics/tehgladiators.py
@@ -10,6 +10,7 @@ class ComicData(ComicDataBase):
 
 class Crawler(CrawlerBase):
     history_capable_days = 90
+    schedule = 'We'
     time_zone = 1
 
     def crawl(self, pub_date):