+++ comics/comics/walkoflife.py
@@ -12,9 +12,11 @@ class ComicData(ComicDataBase):
 
 class Crawler(CrawlerBase):
     history_capable_date = '2008-06-23'
+    schedule = 'Tu,Fr'
     time_zone = 1
 
     def crawl(self, pub_date):
+        epoch = self.date_to_epoch(pub_date, 'Europe/Oslo')
         url = 'http://walkoflife.nettserier.no/_striper/walkoflife-%s.png' % (
-            self.date_to_epoch(pub_date),)
+            epoch,)
         return CrawlerImage(url)