+++ comics/comics/whattheduck.py
@@ -10,7 +10,7 @@ class ComicData(ComicDataBase):
 
 class Crawler(CrawlerBase):
     history_capable_days = 7
-    schedule = 'Mo'
+    schedule = 'We'
 
     def crawl(self, pub_date):
         feed = self.parse_feed('http://www.whattheduck.net/strip/rss.xml')