+++ comics/comics/whiteninja.py
@@ -10,7 +10,6 @@ class ComicData(ComicDataBase):
 
 class Crawler(CrawlerBase):
     history_capable_days = 60
-    schedule = 'Mo,We,Fr'
     time_zone = -6
 
     def crawl(self, pub_date):