+++ comics/comics/yehudamoon.py
@@ -1,5 +1,3 @@
-import re
-
 from comics.aggregator.crawler import CrawlerBase, CrawlerImage
 from comics.core.comic_data import ComicDataBase
 
@@ -12,29 +10,14 @@ class ComicData(ComicDataBase):
 
 class Crawler(CrawlerBase):
     history_capable_date = '2008-01-22'
-    schedule = 'Mo'
+    schedule = 'Mo,Tu,We,Th,Fr'
     time_zone = -5
 
     def crawl(self, pub_date):
-        page_url = 'http://www.yehudamoon.com/index.php?date=%s' % \
-            pub_date.strftime('%Y-%m-%d')
+        page_url = 'http://www.yehudamoon.com/%s/' % (
+            pub_date.strftime('%m%d%Y'),)
         page = self.parse_page(page_url)
-
-        # It'll forward you to the most current day if it doesn't have
-        # pub_date. Check *explicitly* to make sure this day exists and bug out
-        # if not
-        current_day = page.value('select[id=ss_select] option[value*=%s]' %
-            pub_date.strftime('%Y-%m-%d'))
-
-        if current_day is None:
-            return
-
-        url = page.src('div[id="ss_img_div"] img[src*=strips]')
-        # If we can't figure the title out, just don't store it
-        try:
-            title_full = page.text('option[value*="%s"]' %
-                pub_date.strftime('%Y-%m-%d'))
-            title = re.sub('^.*- *', '', title_full)
-        except TypeError:
-            title = None
+        url = page.src('#comic img')
+        title = page.alt('#comic img')
+        title = title.split(' ', 2)[-1]
         return CrawlerImage(url, title)