+++ comics/comics/zelda.py
@@ -10,11 +10,12 @@ class ComicData(ComicDataBase):
 
 class Crawler(CrawlerBase):
     history_capable_days = 30
+    schedule = 'Mo,Tu,We,Th,Fr,Sa'
     time_zone = 1
 
     def crawl(self, pub_date):
-        page_url = 'http://www.dagbladet.no/tegneserie/zelda/?%s' % (
-            self.date_to_epoch(pub_date),)
+        epoch = self.date_to_epoch(pub_date, 'Europe/Oslo')
+        page_url = 'http://www.dagbladet.no/tegneserie/zelda/?%s' % epoch
         page = self.parse_page(page_url)
         url = page.src('img#zelda-stripe')
         return CrawlerImage(url)